// java-components/management-console/src/main/webui/src/services/openapi/models/DependencySetDTO.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { IdentifiedDependencyDTO } from './IdentifiedDependencyDTO';
export type DependencySetDTO = {
    id?: number;
    name?: string;
    dependencies?: Array<IdentifiedDependencyDTO>;
    totalDependencies: number;
    untrustedDependencies: number;
    trustedDependencies: number;
    availableBuilds: number;
};